lsp_normalize: null-terminate returned data
Sporadic bus errors were noticed, while searches were active.

When searching, regexec(3) is used to search in the normalized data of
lines and that data isn't null-terminated, but regexec(3) requires
that data to be null-terminated.

Fix this in lsp_normalize() so that null-terminated data will be
returned.

The term "null-terminated data" is used, because in case of binary
input, that data itself can contain null-characters.

Signed-off-by: Dirk Gouders <dirk@gouders.net>