lsp_file_do_read: also output pre-read byte
When we get fed by a preprocessor, we do an initial read(2) of one
byte on the input.  That byte was still missing in output files that
can be created using option "-o".

Fix this by first writing that byte to the output file before the
then first read block.

Fixes: #31
Signed-off-by: Dirk Gouders <dirk@gouders.net>