lsp: refactor functions that return data bytes
Functions that return bytes that are part of our input return those
bytes as return values -- intermixed with special values for EOF or
beginning of file that could also be valid values for the input data.

Let's separate returned data and result: return the asked for byte via
an argument and leave the return value for just error/success
indication.

Signed-off-by: Dirk Gouders <dirk@gouders.net>