doc: Add trailing newlines to text files
Neither INSTALL nor README.md have trailing newlines.

This may lead to problems, when processing those files. cat(1), for instance,
does not add a trailing newline, if it is missing: The shell prompt will appear
on the same line as the printout of the last line of README.md and INSTALL.

According to the POSIX standard, a line (3.206) is terminated by a newline
character and text files (3.403) contain zero or more lines. Without the
trailing newline, the current last "line" is incomplete -- hence the problems
mentioned above.

Introduced in
- 558568d (lsp: version 0.6.0, 2025-12-06)
- 6275af7 (doc: Add INSTALL instructions., 2024-04-05)

Both
- git show 558568d README.md
- git show 6275af7 INSTALL
notify about the missing newline at the end of the file.

Add newlines to the end of both files.

Signed-off-by: Tom Vierjahn <tom.vierjahn@acm.org>